기능중심 구조로 전체 dir 구조 변경
