--- app/services/stt_feedback_service.py
@@ -1,1 +1,0 @@
-#실제 서비스 로직 작성
